ai-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
spacy==3.7.2
scikit-learn==1.3.2
nltk==3.8.1
textstat==0.7.3
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
